+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [unreleased](https://github.com/mozilla/glam/compare/2020.11.0...HEAD) (date TBD)
+## [unreleased](https://github.com/mozilla/glam/compare/2020.11.1...HEAD) (date TBD)
+
+## [2020.11.1](https://github.com/mozilla/glam/compare/2020.11.0...2020.11.1) (2020-11-12)
 
 - Provide SQL query for further data exploration
   ([#1021](https://github.com/mozilla/glam/pull/1021))